+++ R/reload.R
@@ -119,7 +119,7 @@ document_and_reload <- function(
     )
   })
   if (attempt::is_try_error(roxed)) {
-    cat_rule(
+    cli_cat_rule(
       "Error documenting your package"
     )
     dialog_if_has("Alert", "Error documenting your package")
@@ -136,7 +136,7 @@ document_and_reload <- function(
   })
 
   if (attempt::is_try_error(loaded)) {
-    cat_rule(
+    cli_cat_rule(
       "Error loading your package"
     )
     dialog_if_has("Alert", "Error loading your package")