+++ NEWS.md
@@ -12,7 +12,7 @@
 
 ## New features / user visible changes
 
-+ The following hard dependencies have been moved to soft dependencies, as you'll only need them while developing: `{pkgload}`, `{roxygen2}`, `{fs}`, `{usethis}`
++ The following hard dependencies have been moved to soft dependencies, as you'll only need them while developing: `{pkgload}`, `{roxygen2}`, `{fs}`, `{usethis}`, `{cli}`, `{crayon}`
 + Soft dependency check is now done via `rlang::check_installed()` (#835)
 + `golem::run_dev()` has been refactored to match the behavior of other functions, notably it now uses `golem::get_golem_wd()` to find the current working dir.
 + `{golem}` now depends on `{rlang}` version >= 1.0.0